requirement.txt
django==3.2.18
environs==9.5.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0
django-phonenumber-field==7.0.2
phonenumbers==8.13.4
django_debug_toolbar==3.8.1
django-smart-selects==1.6.0
rosreestr2coord==4.1.6
folium==0.14.0
python-docx==0.8.11